Update weighted scoring to reflect new department rules
Modify backend logic and frontend display to implement new weighted scoring system for GS and Offer departments, including decimal score representation and updated column headers.

+++ backend/schemas.py
@@ -440,7 +440,7 @@ class StaffPerformance(BaseModel):
     gs_count: int
     offer_count: int
     status_breakdown: dict
-    weighted_workload: int = 0   # GS-stage active×3 + CoE/Visa active×2 + Offer active×1
+    weighted_workload: float = 0  # GS core×3 + Visa Lodged×2 + Offer×1 + CoE/Not Eligible×0.1
 
 
 class StaffTimingReport(BaseModel):
